Fix dobot movement for box shape detection - Use MQTT to send the current view from dobot script to the deployment_page to be able to check the labels - Add labels for different views for P2/143 machine part - TO TEST for P2/143 training and running

+++ src/lib/machine_learning/command_utils.py
@@ -376,9 +376,9 @@ def run_command_update_metrics(
         return traceback
 
 
-def export_tfod_savedmodel(training_paths: Dict[str, Path]) -> bool:
+def export_tfod_savedmodel(training_paths: Dict[str, Path], re_export: bool = True) -> bool:
     paths = training_paths
-    if paths['export'].exists():
+    if not re_export and paths['export'].exists():
         # remove any existing export directory first
         shutil.rmtree(paths['export'])
